requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
groq==0.4.2
python-dotenv==1.0.0
notion-client==2.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.2
httpx==0.24.1 